server/requirements.txt
annotated-types>=0.7.0
numpy>=1.26.0
attrs>=25.0.0
bcrypt>=4.0.0
blinker>=1.9.0
cffi>=1.16.0
click>=8.1.0
colorama>=0.4.6
cryptography>=42.0.0
cvat_sdk>=2.48.0
dnspython>=2.6.0
email-validator>=2.0.0
Flask>=3.0.0
Flask-PyMongo>=3.0.0
Flask-Cors>=4.0.0
idna>=3.6
itsdangerous>=2.2.0
Jinja2>=3.1.0
jwt>=1.3.0
MarkupSafe>=3.0.0
packaging>=23.0
pillow>=10.0.0
platformdirs>=4.0.0
pycparser>=2.21
pydantic>=2.5.0
pydantic_core>=2.14.0
pymongo>=4.6.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
setuptools>=68.0.0
six>=1.16.0
tqdm>=4.65.0
typing-inspection>=0.4.0
typing_extensions>=4.8.0
urllib3>=2.0.0
Werkzeug>=3.0.0
cellpose

ultralytics
opencv-python-headless
matplotlib
torch
